requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
scikit-learn==1.7.2
numpy==1.24.3
pyyaml==6.0.1
mlflow==2.9.2

# Authentication dependencies
sqlalchemy==2.0.25
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.3
